refactor(agents): extract Runtime/Model/ReviewPolicy pickers into shared modules
Splits each picker into a bare *Select (for table cells) and a card-wrapped
*Picker (for the detail aside) so the upcoming list view can reuse the same
mutation + select logic without dragging the section chrome into table rows.

No behavior change to /agents/[id] — the aside still renders the same three
cards. Re-exports AgentDisplay from the web types so pickers can be typed
against the lighter shape used by both detail and list contexts.
